+++ lib/unisec/cli/cli.rb
@@ -4,6 +4,8 @@ require 'unisec/cli/surrogates'
 require 'unisec/cli/hexdump'
 require 'unisec/cli/properties'
 require 'unisec/cli/confusables'
+require 'unisec/cli/versions'
+require 'unisec/cli/size'
 
 module Unisec
   # Module used to create the CLI for the executable
@@ -22,6 +24,8 @@ module Unisec
       register 'properties char', Properties::Char
       register 'confusables list', Confusables::List
       register 'confusables randomize', Confusables::Randomize
+      register 'versions', Versions
+      register 'size', Size
     end
   end
 end