+++ docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## [unreleased]
 
+## [0.0.2]
+
+- Add 2 new classes (and corresponding CLI command):
+  - `Unisec::Versions`: Version of Unicode, ICU, CLDR, gems used in Unisec
+  - `Unisec::Size`: Code point, grapheme, UTF-8/UTF-16/UTF-32 byte/unit size
+
 ## [0.0.1]
 
 - Initial version